requirements/dev.txt
-r base.txt

jedi
flake8
django-debug-toolbar
pytest
pytest-django
factory_boy
